Updates variable definitions to more confusion verbiage
In an attempt to make things clearer I've upadated all the variable
definitons to use more words, making it tedious and more difficult to
read and understand to people outside of the coding industry because
evidently the desire of coders is to make sure no one understands what
the are actually saying when they attempt to communicate with people.
But hey, if that's what my superiors want, that's what I'll do.

[#96449482]