Adds lines 11 and 17 to prove @total changes
It was said that @total_owed would not change if @total_owed =+
amount_owed was used to define the purchase method.  To disprove this a
line was added directly underneath this to the the purchase method that
shows the new @total_owed when the purchase parameter (amount_owed) is
set.  It reads as follows: puts "You owe $X.XX" and the X.XX taken from the new
@total_owed.

[#96738650]